Convert export task models to use common base class
